Move Favorites list handlers to the relevant modules and order dashboard favorites
